# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cafrezzo REST API
  version: "1.0.0"
  description: |
    Backend REST API for **Cafrezzo** — premium coffee e-commerce platform.

    ### Base URL
    - Local: `http://localhost:8000/api/v1`
    - Production: `https://api.cafrezzo.com/api/v1`

    ### Authentication
    Most endpoints are public. Authenticated endpoints require a **Bearer token**
    obtained from `POST /auth/login`. Pass it as: `Authorization: Bearer {token}`

    ### Localization
    Add `Accept-Language: fr|en|de|ru|nl` header for localized responses.
  contact:
    email: dev@cafrezzo.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.cafrezzo.com/api/v1
    description: Production
components:
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: Token from POST /auth/login
...
paths:
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Send password reset link by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200": { description: Reset link sent }
        "422": { description: Email not found }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using token from email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, email, password, password_confirmation]
              properties:
                token:                 { type: string }
                email:                 { type: string, format: email }
                password:              { type: string, minLength: 8 }
                password_confirmation: { type: string }
      responses:
        "200": { description: Password reset success }
        "422": { description: Invalid token }

  # ── ORDERS (authenticated) ────────────────────────────────────
...